增加rabbitmq增量同步用户数据功能

+++ app/src/main/java/net/ischool/isus/model/Uids.kt
@@ -9,4 +9,4 @@ import com.google.gson.annotations.SerializedName
  * Email: zhaocework@gmail.com
  * Date: 2019-05-30
  */
-data class Uids(@SerializedName(value = "list") val uids: List<Int>)
+data class Uids(@SerializedName(value = "list") val uids: List<Long>)
